[#113] Use the price item field type for rate plan price fields.
